annotator/requirements.txt
psycopg2
python-dotenv
srsly==2.4.8